--- src/app/header.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-header',
-  template: `
-    <p>
-      header works!
-    </p>
-  `,
-  styles: []
-})
-export class HeaderComponent implements OnInit {
-  constructor() {}
-  ngOnInit() {}
-}
